p4i integration Integrate 169634 from 4.0 to main: store the font size internally as a qreal, not as deci-point
see task 61111

reviewed by lars

[git-p4: depot-paths = "//depot/qt/main/": change = 169635]

+++ src/gui/text/qfont_p.h
@@ -36,7 +36,7 @@ class QFontEngine;
 struct QFontDef
 {
     inline QFontDef()
-        : pointSize(-1), pixelSize(-1),
+        : pointSize(-1.0), pixelSize(-1),
           styleStrategy(QFont::PreferDefault), styleHint(QFont::AnyStyle),
           weight(50), fixedPitch(false), style(QFont::StyleNormal), stretch(100),
           ignorePitch(true)
@@ -52,7 +52,7 @@ struct QFontDef
     QString addStyle;
 #endif // Q_WS_X11
 
-    int pointSize;
+    qreal pointSize;
     int pixelSize;
 
     uint styleStrategy : 16;
